+++ src/main/kotlin/org/olafneumann/mahjong/points/ui/html/Bootstrap.kt
@@ -53,58 +53,3 @@ fun TagConsumer<HTMLElement>.closeButton(
         tooltip?.let { this.title = !it }
         this.onClickFunction = onClickFunction
     }
-
-fun TagConsumer<HTMLElement>.modal2(
-    title: String,
-    buttons: List<Button> = emptyList(),
-    mainBlock: TagConsumer<HTMLElement>.() -> Unit = {}
-) =
-    injectRoot { element ->
-        element.addEventListener("hidden.bs.modal", { element.remove() })
-    }
-        .div(classes = "modal fade") {
-            div(classes = "modal-dialog modal-dialog-centered") {
-                div(classes = "modal-content") {
-                    div(classes = "modal-header") {
-                        h5(classes = "modal-title") { +!title }
-                        closeButton(additionalAttributes = listOf("data-bs-dismiss" to "modal"))
-                    }
-                    div(classes = "modal-body") {
-                        mainBlock()
-                    }
-                    div(classes = "modal-footer") {
-                        buttons.forEach {
-                            bsButton(label = it.title, colorClass = it.colorClass, onClickFunction = it.onClickFunction)
-                        }
-                    }
-                }
-            }
-        }
-
-data class Button(
-    val title: String,
-    val colorClass: String = "primary",
-    val onClickFunction: (Event) -> Unit
-)
-
-@Suppress("UnusedPrivateMember") // both members are used in JS code
-fun createModal(element: HTMLElement, static: Boolean = true): Modal {
-    val options = createOptionsJson(static = static)
-    val modal = js("new bootstrap.Modal(element, options)")
-    return modal.unsafeCast<Modal>()
-}
-
-private fun createOptionsJson(static: Boolean) =
-    mapOf(
-        "backdrop" to if (static) "static" else "",
-        "focus" to true,
-        "keyboard" to true
-    )
-        .toJson()
-
-
-external class Modal {
-    fun show()
-    fun hide()
-}
-
